refactor(tests): split retry budget tests into separate file to meet LOC threshold
- Extract TestRetryBudget class to test_retry_budget.py
- Reduce test_dispatch_state_transitions.py from 629 to 316 lines
- Both files now under 400-line CI threshold
- All 153 orchestra tests pass

Refs: #774
